Feature/implement data related logic for ranges (#2)
* Feature: Implement data-related logic for RangeData and update range factory methods for consistency

- Introduced RangeData abstraction to couple logical ranges with data sequences and domains.
- Updated RangeFactory methods to remove type parameters for cleaner usage.
- Added TryCreate method for Range with validation and error messaging.
- Enhanced Distance method in domain interfaces to ensure accurate step calculations.
- Revised tests to reflect changes in range factory methods and ensure correctness.
- Updated project metadata and CI/CD configurations for new data module.

* Feature: Add validation method for RangeData to ensure data sequence matches expected logical elements

* Feature: Add overview and implementation details for RangeData in README.md

* Feature: Add unit tests for RangeData extensions to validate adjacency and data integrity

* Feature: Add unit tests for RangeData extensions to validate equality, slicing, and domain interactions

* Feature: Add benchmarks for RangeData and RangeDataExtensions to evaluate performance and memory usage

* Feature: Integrate Codecov coverage reporting into CI workflows for all projects

* Feature: Enhance RangeData logic for sub-range and point access with validation and improved error handling

- Refine RangeData indexer to validate sub-range finiteness and containment, throwing specific exceptions for invalid cases
- Update TryGet methods to ensure points and sub-ranges are within bounds before accessing data
- Align index calculations with range inclusivity for accurate data mapping
- Improve XML documentation for IRangeDomain to clarify complexity of Distance implementations
- Update benchmarks and tests for readonly domain fields and minor cleanup
- Clarify package description and README section for RangeData usage and validation

* Feature: Update Microsoft.NET.Test.Sdk to version 17.11.1 in test project dependencies

* Feature: Update Microsoft.NET.Test.Sdk to version 17.11.1 in test project dependencies

* Feature: Update test execution paths and remove Codecov upload from Domain.Abstractions workflow

* Feature: Update test paths in CI configuration and enhance documentation for TryCreate method

* Feature: Update CI configuration to use test paths for dependency restoration, building, and testing

* Feature: Correct test path in CI configuration for Intervals.NET.Data.Tests

* Feature: Update CI configuration to use test path for dependency restoration and building

* Feature: Optimize null-check for domain parameter to avoid boxing in RangeData

* Feature: Enhance test coverage for RangeData and DateTime domains with edge case scenarios

* chore: get rid of redundant md files

* Feature: Update range comments and improve test method naming for clarity

* Feature: Refine distance calculation and improve exception messages in RangeData

* Feature: Improve range validation logic to handle overflow and enhance error messaging

* Feature: Enhance range validation to correctly handle infinity bounds and improve exception messaging

* Feature: Update range documentation for clarity and improve validation logic in TrimEnd method

* Feature: Enhance TrimStart and TrimEnd methods to support inclusive/exclusive boundaries and improve documentation

* Feature: Update documentation for RangeData methods to clarify return values and equality semantics

* Feature: Clarify documentation for range span calculation and return type details

* Feature: Update test data in IsTouching method to clarify range element counts

* Feature: Update test data in IsBeforeAndAdjacentTo method to clarify element counts for ranges

* Feature: Refine documentation and logic in RangeData extensions for clarity and consistency

* Feature: Enhance documentation for domain interfaces and implementations to clarify performance characteristics and equality semantics

* Feature: Update README to enhance structure and clarity of RangeData documentation

* Fix TryGet returning false for empty subranges with exclusive parent boundaries (#3)

* Initial plan

* Add test case demonstrating bug with empty subranges in exclusive-start parents


* Fix bug where empty subranges with negative indices returned false instead of empty RangeData


---------


---------

+++ src/Domain/Intervals.NET.Domain.Extensions/Variable/RangeDomainExtensions.cs
@@ -79,13 +79,15 @@ public static class RangeDomainExtensions
     /// <typeparam name="TRangeValue">The type of the values in the range. Must implement IComparable&lt;T&gt;.</typeparam>
     /// <typeparam name="TDomain">The type of the domain that implements IVariableStepDomain&lt;TRangeValue&gt;.</typeparam>
     /// <returns>
-    /// The span (distance) of the range as a double, potentially including fractional steps,
-    /// or infinity if the range is unbounded.
+    /// The span (distance) of the range as a double representing the number of complete discrete domain steps,
+    /// or positive infinity if the range is unbounded.
+    /// The return type is double to accommodate infinity values; the actual step count is always an integer (converted from long).
     /// </returns>
     /// <remarks>
     /// <para>
-    /// Counts the number of domain steps that fall within the range boundaries, respecting inclusivity.
-    /// Unlike fixed-step domains, this may return fractional values to account for partial steps.
+    /// Counts the number of discrete domain steps that fall within the range boundaries, respecting inclusivity.
+    /// Variable-step domains use <see cref="IRangeDomain{T}.Distance"/> which returns a long (discrete step count),
+    /// not fractional distances. The double return type exists solely to represent infinity for unbounded ranges.
     /// </para>
     /// 
     /// <para>
@@ -122,7 +124,9 @@ public static class RangeDomainExtensions
                 return HandleSingleStepCase(range, domain);
         }
 
-        var distance = domain.Distance(firstStep, lastStep);
+        // Note: IRangeDomain.Distance returns long, but for variable-step domains we interpret
+        // this as the number of complete steps and add 1.0 to get the span including boundaries
+        var distance = (double)domain.Distance(firstStep, lastStep);
         return distance + 1.0;
 
         // Local functions
